Overhaul of Scylla Engine, replace with GridEngine
This is a major overhual of networking infrastructure. The ScyllaEngine
is now deprecated and replaced with new GridEngine interface (also note
that the namespace is also change to zillians::networking). We will add
more transport type (spread/session/infiniband) to GridEngine.

Also we provide a simple test case GridEngineBasicTest (inside the new
directory testzillians-common-network/GridEngineBasicTest)

The current version of CPG transport in GridEngine does not support
multiple ack on a single queue (destination) now, but will be fixed
later.

	Testing done: Yes
	Reviewer: N/A